+++ src/Symfony/Component/ObjectMapper/TransformCallableWithContextInterface.php
@@ -19,12 +19,13 @@ namespace Symfony\Component\ObjectMapper;
  *
  * {@see Symfony\Component\ObjectMapper\Attribute\Map}
  */
-interface TransformCallableInterface
+interface TransformCallableWithContextInterface
 {
     /**
      * @param mixed   $value   The value being mapped
      * @param T       $source  The object we're working on
      * @param T2|null $target  The target we're mapping to
+     * @param array   $context The mapping context
      */
-    public function __invoke(mixed $value, object $source, ?object $target): mixed;
+    public function __invoke(mixed $value, object $source, ?object $target, array $context): mixed;
 }